Refatoração de código

+++ Services/ApiOne/Domain/Interfaces/ITaxaJurosDomainService.cs
@@ -2,6 +2,6 @@
 {
     public interface ITaxaJurosDomainService
     {
-        double ObterTaxaDeJuros();
+        decimal ObterTaxaDeJuros();
     }
 }
