+++ Services/ApiOne/Domain/Services/TaxaJurosDomainService.cs
@@ -4,9 +4,9 @@ namespace ApiOne.Domain.Services
 {
     public class TaxaJurosDomainService : ITaxaJurosDomainService
     {
-        public double ObterTaxaDeJuros()
+        public decimal ObterTaxaDeJuros()
         {
-            var taxaDeJuros = 0.01;
+            var taxaDeJuros = 0.01m;
             return taxaDeJuros;
         }
     }
